requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
auth==0.5.3
bcrypt==3.2.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
eventlet==0.39.1
falcon==4.0.2
fastapi==0.115.12
frozenlist==1.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.0
grpcio==1.71.0
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
lockfile==0.12.2
Mako==1.3.10
MarkupSafe==3.0.2
mongoengine==0.29.1
multidict==6.4.3
openai==1.71.0
packaging==25.0
passlib==1.7.4
pid==3.0.4
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
pymongo==4.12.0
pyparsing==3.2.3
python-daemon==3.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
repoze.lru==0.7
requests==2.32.3
Routes==2.5.1
rsa==4.9
service==0.6.0
services==0.1.1
setproctitle==1.3.5
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.1
tqdm==4.67.1
twilio==9.6.1
typing-inspection==0.4.0
typing_extensions==4.13.1
uritemplate==4.1.1
urllib3==2.4.0
uvicorn==0.34.0
Werkzeug==3.1.3
yarl==1.20.0
